+++ src/ServerEntry.cpp
@@ -1,11 +1,11 @@
 //
 // ServerEntry.cpp for raytracer in /home/jochau_g//Desktop/raytracer-42/src
-// 
+//
 // Made by gael jochaud-du-plessix
 // Login   <jochau_g@epitech.net>
-// 
+//
 // Started on  Thu May 26 18:17:38 2011 gael jochaud-du-plessix
-// Last update Wed Jun  1 14:03:39 2011 gael jochaud-du-plessix
+// Last update Wed Jun  1 21:20:47 2011 melvin laplanche
 //
 
 #include <sstream>
@@ -19,7 +19,7 @@
 ServerEntry::ServerEntry():
   _clusterClient(NULL), _ip(), _port(0), _status(0), _progress(0),
   _socket(NULL), _currentRequest(-1), _currentPacketSize(0)
-{  
+{
 }
 
 ServerEntry::ServerEntry(ClusterClient* clusterClient, QString ip, int port,
@@ -27,7 +27,7 @@ ServerEntry::ServerEntry(ClusterClient* clusterClient, QString ip, int port,
   _clusterClient(clusterClient), _ip(ip), _port(port), _status(status),
   _progress(progress), _socket(NULL), _currentRequest(-1),
   _currentPacketSize(0)
-{  
+{
 }
 
 ServerEntry::~ServerEntry()
@@ -89,7 +89,7 @@ void		ServerEntry::openConnection(void)
   if (isConnectionOpened())
     return ;
   if (!_socket)
-    {      
+    {
       _socket = new QTcpSocket();
       connect(_socket, SIGNAL(connected()), this,
 	      SLOT(onConnectionOpened()));
@@ -151,7 +151,7 @@ void		ServerEntry::onDataReceived(void)
   QDataStream	stream(_socket);
   if (_currentRequest == -1)
     {
-      if (_socket->bytesAvailable() < sizeof(int))
+      if ((size_t)_socket->bytesAvailable() < sizeof(int))
 	return ;
       stream >> _currentRequest;
     }
@@ -189,7 +189,7 @@ void		ServerEntry::onDataReceived(void)
     {
       if (_currentPacketSize == 0)
 	{
-	  if (_socket->bytesAvailable() < sizeof(int))
+	  if ((size_t)_socket->bytesAvailable() < sizeof(int))
 	    return ;
 	  stream >> _currentPacketSize;
 	}
